+++ src/components/themes/BaseTheme.h
@@ -64,7 +64,7 @@ struct ThemeMetrics {
   bool keyboardCenteredText;
 };
 
-enum UIIcon { Folder, Text, Image, Book, File, Recent, Settings, Transfer, Library, Wifi, Hotspot, Todo };
+enum UIIcon { Folder, Text, Image, Book, File, Recent, Settings, Transfer, Library, Wifi, Hotspot, Todo, Cal };
 
 // Default theme implementation (Classic Theme)
 // Additional themes can inherit from this and override methods as needed